requirements.txt
pandas
tqdm
spacy
praw
python-dotenv
requests
pydeidentify
newsapi-python
 beautifulsoup4
lxml
selenium
webdriver-manager
openai-whisper
torch
